fixed films

+++ cinemaManager/views/film.py
@@ -1,6 +1,6 @@
 from django.shortcuts import render, redirect
 from ..forms.FilmForm import FilmForm
-from ..models.general import Film 
+from ..models.general import Film, Showing
 
 def addFilmForm(request):
     form = FilmForm()
@@ -20,11 +20,19 @@ def deleteFilm(request):
     if request.method == 'POST':
         film_id = request.POST.get('film_id')
         if film_id:
-            Film.objects.filter(film_id=film_id).delete()
-            return redirect('filmList')
+            # Check if there are any showings associated with this film
+            num_showings = Showing.objects.filter(film_id=film_id).count()
+            if num_showings == 0:
+                Film.objects.filter(film_id=film_id).delete()
+                return redirect('filmList')
+            else:
+                # There are showings associated with this film, so don't delete it
+                films = Film.objects.all()
+                context = {'films': films, 'error': 'Cannot delete film with associated showings.'}
+                return render(request, 'Films/DeleteFilm.html', context)
     films = Film.objects.all()
     context = {'films': films}
-    return render(request, 'Films/DeleteFilm.html', context)   
+    return render(request, 'Films/DeleteFilm.html', context)
 
 def filmList(request):
     films = Film.objects.all()
